+++ packages/ruleset/src/functions/index.js
@@ -4,8 +4,10 @@ const checkMajorVersion = require('./check-major-version');
 const descriptionMentionsJSON = require('./description-mentions-json');
 const disallowedHeaderParameter = require('./disallowed-header-parameter');
 const discriminator = require('./discriminator');
+const duplicatePathParameter = require('./duplicate-path-parameter');
 const enumCaseConvention = require('./enum-case-convention');
 const errorResponseSchema = require('./error-response-schema');
+const inlineResponseSchema = require('./inline-response-schema');
 const operationIdCaseConvention = require('./operation-id-case-convention');
 const operationIdNamingConvention = require('./operation-id-naming-convention');
 const operationSummary = require('./operation-summary');
@@ -14,6 +16,7 @@ const parameterCaseConvention = require('./parameter-case-convention');
 const parameterDefault = require('./parameter-default');
 const parameterDescription = require('./parameter-description');
 const parameterOrder = require('./parameter-order');
+const pathSegmentCaseConvention = require('./path-segment-case-convention');
 const propertyCaseCollision = require('./property-case-collision');
 const propertyCaseConvention = require('./property-case-convention');
 const propertyDescription = require('./property-description');
@@ -21,6 +24,7 @@ const propertyInconsistentNameAndType = require('./property-inconsistent-name-an
 const requestBodyName = require('./request-body-name');
 const requiredProperty = require('./required-property');
 const responseExampleProvided = require('./response-example-provided');
+const responseStatusCodes = require('./response-status-codes');
 const schemaDescription = require('./schema-description.js');
 const schemaOrContentProvided = require('./schema-or-content-provided');
 const { securitySchemes } = require('./security-schemes');
@@ -35,8 +39,10 @@ module.exports = {
   checkMajorVersion,
   descriptionMentionsJSON,
   discriminator,
+  duplicatePathParameter,
   enumCaseConvention,
   errorResponseSchema,
+  inlineResponseSchema,
   operationIdCaseConvention,
   operationIdNamingConvention,
   operationSummary,
@@ -45,6 +51,7 @@ module.exports = {
   parameterDefault,
   parameterDescription,
   parameterOrder,
+  pathSegmentCaseConvention,
   propertyCaseCollision,
   propertyCaseConvention,
   propertyDescription,
@@ -52,6 +59,7 @@ module.exports = {
   requestBodyName,
   requiredProperty,
   responseExampleProvided,
+  responseStatusCodes,
   schemaDescription,
   schemaOrContentProvided,
   stringBoundary,