+++ packages/ruleset/src/rules/index.js
@@ -7,11 +7,13 @@ const contentEntryProvided = require('./content-entry-provided');
 const contentTypeParameter = require('./content-type-parameter');
 const descriptionMentionsJSON = require('./description-mentions-json');
 const discriminator = require('./discriminator');
+const duplicatePathParameter = require('./duplicate-path-parameter');
 const enumCaseConvention = require('./enum-case-convention');
 const examplesNameContainsSpace = require('./examples-name-contains-space');
 const ibmContentTypeIsSpecific = require('./ibm-content-type-is-specific');
 const ibmErrorContentTypeIsJson = require('./ibm-error-content-type-is-json');
 const ibmSdkOperations = require('./ibm-sdk-operations');
+const inlineResponseSchema = require('./inline-response-schema');
 const majorVersionInPath = require('./major-version-in-path');
 const missingRequiredProperty = require('./missing-required-property');
 const operationIdCaseConvention = require('./operation-id-case-convention');
@@ -23,6 +25,7 @@ const parameterDefault = require('./parameter-default');
 const parameterDescription = require('./parameter-description');
 const parameterOrder = require('./parameter-order');
 const parameterSchemaOrContent = require('./parameter-schema-or-content');
+const pathSegmentCaseConvention = require('./path-segment-case-convention');
 const prohibitSummarySentenceStyle = require('./prohibit-summary-sentence-style');
 const propertyCaseCollision = require('./property-case-collision');
 const propertyCaseConvention = require('./property-case-convention');
@@ -32,6 +35,7 @@ const requestBodyName = require('./request-body-name');
 const requestBodyObject = require('./request-body-object');
 const responseErrorResponseSchema = require('./response-error-response-schema');
 const responseExampleProvided = require('./response-example-provided');
+const responseStatusCodes = require('./response-status-codes');
 const schemaDescription = require('./schema-description');
 const securitySchemes = require('./security-schemes');
 const serverVariableDefaultValue = require('./server-variable-default-value');
@@ -49,11 +53,13 @@ module.exports = {
   contentTypeParameter,
   descriptionMentionsJSON,
   discriminator,
+  duplicatePathParameter,
   enumCaseConvention,
   examplesNameContainsSpace,
   ibmContentTypeIsSpecific,
   ibmErrorContentTypeIsJson,
   ibmSdkOperations,
+  inlineResponseSchema,
   majorVersionInPath,
   missingRequiredProperty,
   operationIdCaseConvention,
@@ -65,6 +71,7 @@ module.exports = {
   parameterDescription,
   parameterOrder,
   parameterSchemaOrContent,
+  pathSegmentCaseConvention,
   prohibitSummarySentenceStyle,
   propertyCaseCollision,
   propertyCaseConvention,
@@ -74,6 +81,7 @@ module.exports = {
   requestBodyObject,
   responseErrorResponseSchema,
   responseExampleProvided,
+  responseStatusCodes,
   schemaDescription,
   securitySchemes,
   serverVariableDefaultValue,